third_party/WebKit/LayoutTests/media/mediasession/mediasessionmetadata.html

<!DOCTYPE html>
<title>MediaMetadata interface</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

  test(function() {
    var metadata = new MediaMetadata({});
    assert_not_equals(metadata, null);

    var exception = false;
    try {
      metadata = new MediaMetadata("foobar");
    } catch (e) {
      exception = true;
    }
    assert_true(exception);

    exception = false;
    try {
      metadata = new MediaMetadata(42);
    } catch (e) {
      exception = true;
  }
    assert_true(exception);
  }, 'Test that MediaMetadata is constructed using a dictionary');

  test(function() {
    var metadata = new MediaMetadata({
        title: 'foo', album: 'bar', artist: 'plop',
        artwork: [ { src: 'src1', sizes: 'sizes1', type: 'type1'},
                   { src: 'src2', sizes: 'sizes2', type: 'type2'} ] });
    assert_equals(metadata.title, 'foo');
    assert_equals(metadata.album, 'bar');
    assert_equals(metadata.artist, 'plop');
    assert_equals(2, metadata.artwork.length);

    var image1 = new MediaImage({ src: 'src1', sizes: 'sizes1', type: 'type1'});
    var image2 = new MediaImage({ src: 'src2', sizes: 'sizes2', type: 'type2'});
    assert_equals(metadata.artwork[0].src, image1.src);
    assert_equals(metadata.artwork[0].sizes, image1.sizes);
    assert_equals(metadata.artwork[0].type, image1.type);
    assert_equals(metadata.artwork[1].src, image2.src);
    assert_equals(metadata.artwork[1].sizes, image2.sizes);
    assert_equals(metadata.artwork[1].type, image2.type);
  }, 'Test the different values allowed in MediaMetadata init dictionary');

  test(function() {
    var metadata = new MediaMetadata({});
    assert_equals(metadata.title, '');
    assert_equals(metadata.artist, '');
    assert_equals(metadata.album, '');
    assert_equals(0, metadata.artwork.length);
  }, 'Test the default values for MediaMetadata');

  test(function() {
    var metadata = new MediaMetadata({ randomValueThatWillNotBeAdded: '... hopefully ;)' });
    assert_equals(metadata.randomValueThatWillNotBeAdded, undefined);
  }, 'Test that passing unknown values to the dictionary is a no-op');

</script>
